Camera Follow Code & Some Audio
Wasn't able to do the audio for the Casting Bar script because there wasn't an object with it. I don't know how exactly it was placed in the scene, so I think we can fix this during lab. Also attempted to fix slow movement by making player speed 100. (BTW, one of the sprites for the animation of walking down is the old alpha one. Will also fix in lab)