Debugged Fishing Minigame Issue
Bug 1: "pause" was being called in Lose() function when it shouldn't have been. This was causing the minigame to freeze after a player lost a fish.

Bug 2: Fail timer must be reset to 10 seconds in Win() and Lose() functions. After fixing the pause issue, playing again after losing once would make player lose automatically. This is because the fail timer was not reset and so the game started with -0.0037 seconds, causing it to call Lose().